+++ httprunner/builtin/functions.py
@@ -7,7 +7,6 @@ import random
 import string
 import time
 
-from httprunner.compat import builtin_str, integer_types
 from httprunner.exceptions import ParamsError
 
 
@@ -21,8 +20,8 @@ def gen_random_string(str_len):
 def get_timestamp(str_len=13):
     """ get timestamp string, length can only between 0 and 16
     """
-    if isinstance(str_len, integer_types) and 0 < str_len < 17:
-        return builtin_str(time.time()).replace(".", "")[:str_len]
+    if isinstance(str_len, int) and 0 < str_len < 17:
+        return str(time.time()).replace(".", "")[:str_len]
 
     raise ParamsError("timestamp length can only between 0 and 16.")