+++ httprunner/response.py
@@ -1,10 +1,11 @@
+import json
 import re
 from collections import OrderedDict
 
 import jsonpath
+from loguru import logger
 
-from httprunner import exceptions, logger, utils
-from httprunner.compat import basestring, is_py2
+from httprunner import exceptions, utils
 
 text_extractor_regexp_compile = re.compile(r".*\(.*\).*")
 
@@ -32,39 +33,60 @@ class ResponseObject(object):
             self.__dict__[key] = value
             return value
         except AttributeError:
-            err_msg = "ResponseObject does not have attribute: {}".format(key)
-            logger.log_error(err_msg)
+            err_msg = f"ResponseObject does not have attribute: {key}"
+            logger.error(err_msg)
             raise exceptions.ParamsError(err_msg)
 
-    def _extract_field_with_jsonpath(self, field):
-        """
+    def _extract_field_with_jsonpath(self, field: str) -> list:
+        """ extract field from response content with jsonpath expression.
         JSONPath Docs: https://goessner.net/articles/JsonPath/
-        For example, response body like below:
-        {
-            "code": 200,
-            "data": {
-                "items": [{
-                        "id": 1,
-                        "name": "Bob"
-                    },
-                    {
-                        "id": 2,
-                        "name": "James"
-                    }
-                ]
-            },
-            "message": "success"
-        }
 
-        :param field:  Jsonpath expression, e.g. 1)$.code   2) $..items.*.id
-        :return:       A list that extracted from json repsonse example.    1) [200]   2) [1, 2]
+        Args:
+            field: jsonpath expression, e.g. $.code, $..items.*.id
+
+        Returns:
+            A list that extracted from json response example. 1) [200] 2) [1, 2]
+
+        Raises:
+            exceptions.ExtractFailure: If no content matched with jsonpath expression.
+
+        Examples:
+            For example, response body like below:
+            {
+                "code": 200,
+                "data": {
+                    "items": [{
+                            "id": 1,
+                            "name": "Bob"
+                        },
+                        {
+                            "id": 2,
+                            "name": "James"
+                        }
+                    ]
+                },
+                "message": "success"
+            }
+
+            >>> _extract_field_with_regex("$.code")
+            [200]
+            >>> _extract_field_with_regex("$..items.*.id")
+            [1, 2]
+
         """
-        result = jsonpath.jsonpath(self.parsed_body(), field)
-        if result:
+        try:
+            json_body = self.json
+            assert json_body
+
+            result = jsonpath.jsonpath(json_body, field)
+            assert result
             return result
-        else:
-            raise exceptions.ExtractFailure("\tjsonpath {} get nothing\n".format(field))
-            
+        except (AssertionError, exceptions.JSONDecodeError):
+            err_msg = f"Failed to extract data with jsonpath! => {field}\n"
+            err_msg += f"response body: {self.text}\n"
+            logger.error(err_msg)
+            raise exceptions.ExtractFailure(err_msg)
+
     def _extract_field_with_regex(self, field):
         """ extract field from response content with regex.
             requests.Response body could be json or html text.
@@ -87,9 +109,9 @@ class ResponseObject(object):
         """
         matched = re.search(field, self.text)
         if not matched:
-            err_msg = u"Failed to extract data with regex! => {}\n".format(field)
-            err_msg += u"response body: {}\n".format(self.text)
-            logger.log_error(err_msg)
+            err_msg = f"Failed to extract data with regex! => {field}\n"
+            err_msg += f"response body: {self.text}\n"
+            logger.error(err_msg)
             raise exceptions.ExtractFailure(err_msg)
 
         return matched.group(1)
@@ -120,8 +142,8 @@ class ResponseObject(object):
         if top_query in ["status_code", "encoding", "ok", "reason", "url"]:
             if sub_query:
                 # status_code.XX
-                err_msg = u"Failed to extract: {}\n".format(field)
-                logger.log_error(err_msg)
+                err_msg = f"Failed to extract: {field}\n"
+                logger.error(err_msg)
                 raise exceptions.ParamsError(err_msg)
 
             return getattr(self, top_query)
@@ -136,27 +158,27 @@ class ResponseObject(object):
             try:
                 return cookies[sub_query]
             except KeyError:
-                err_msg = u"Failed to extract cookie! => {}\n".format(field)
-                err_msg += u"response cookies: {}\n".format(cookies)
-                logger.log_error(err_msg)
+                err_msg = f"Failed to extract cookie! => {field}\n"
+                err_msg += f"response cookies: {cookies}\n"
+                logger.error(err_msg)
                 raise exceptions.ExtractFailure(err_msg)
 
         # elapsed
         elif top_query == "elapsed":
             available_attributes = u"available attributes: days, seconds, microseconds, total_seconds"
             if not sub_query:
-                err_msg = u"elapsed is datetime.timedelta instance, attribute should also be specified!\n"
+                err_msg = "elapsed is datetime.timedelta instance, attribute should also be specified!\n"
                 err_msg += available_attributes
-                logger.log_error(err_msg)
+                logger.error(err_msg)
                 raise exceptions.ParamsError(err_msg)
             elif sub_query in ["days", "seconds", "microseconds"]:
                 return getattr(self.elapsed, sub_query)
             elif sub_query == "total_seconds":
                 return self.elapsed.total_seconds()
             else:
-                err_msg = "{} is not valid datetime.timedelta attribute.\n".format(sub_query)
+                err_msg = f"{sub_query} is not valid datetime.timedelta attribute.\n"
                 err_msg += available_attributes
-                logger.log_error(err_msg)
+                logger.error(err_msg)
                 raise exceptions.ParamsError(err_msg)
 
         # headers
@@ -169,16 +191,16 @@ class ResponseObject(object):
             try:
                 return headers[sub_query]
             except KeyError:
-                err_msg = u"Failed to extract header! => {}\n".format(field)
-                err_msg += u"response headers: {}\n".format(headers)
-                logger.log_error(err_msg)
+                err_msg = f"Failed to extract header! => {field}\n"
+                err_msg += f"response headers: {headers}\n"
+                logger.error(err_msg)
                 raise exceptions.ExtractFailure(err_msg)
 
         # response body
         elif top_query in ["body", "content", "text", "json"]:
             try:
                 body = self.json
-            except exceptions.JSONDecodeError:
+            except json.JSONDecodeError:
                 body = self.text
 
             if not sub_query:
@@ -193,9 +215,9 @@ class ResponseObject(object):
                 return utils.query_json(body, sub_query)
             else:
                 # content = "<html>abcdefg</html>", content.xxx
-                err_msg = u"Failed to extract attribute from response body! => {}\n".format(field)
-                err_msg += u"response body: {}\n".format(body)
-                logger.log_error(err_msg)
+                err_msg = f"Failed to extract attribute from response body! => {field}\n"
+                err_msg += f"response body: {body}\n"
+                logger.error(err_msg)
                 raise exceptions.ExtractFailure(err_msg)
 
         # new set response attributes in teardown_hooks
@@ -214,30 +236,30 @@ class ResponseObject(object):
                 return utils.query_json(attributes, sub_query)
             else:
                 # content = "attributes.new_attribute_not_exist"
-                err_msg = u"Failed to extract cumstom set attribute from teardown hooks! => {}\n".format(field)
-                err_msg += u"response set attributes: {}\n".format(attributes)
-                logger.log_error(err_msg)
+                err_msg = f"Failed to extract cumstom set attribute from teardown hooks! => {field}\n"
+                err_msg += f"response set attributes: {attributes}\n"
+                logger.error(err_msg)
                 raise exceptions.TeardownHooksFailure(err_msg)
 
         # others
         else:
-            err_msg = u"Failed to extract attribute from response! => {}\n".format(field)
-            err_msg += u"available response attributes: status_code, cookies, elapsed, headers, content, " \
-                       u"text, json, encoding, ok, reason, url.\n\n"
-            err_msg += u"If you want to set attribute in teardown_hooks, take the following example as reference:\n"
-            err_msg += u"response.new_attribute = 'new_attribute_value'\n"
-            logger.log_error(err_msg)
+            err_msg = f"Failed to extract attribute from response! => {field}\n"
+            err_msg += "available response attributes: status_code, cookies, elapsed, headers, content, " \
+                       "text, json, encoding, ok, reason, url.\n\n"
+            err_msg += "If you want to set attribute in teardown_hooks, take the following example as reference:\n"
+            err_msg += "response.new_attribute = 'new_attribute_value'\n"
+            logger.error(err_msg)
             raise exceptions.ParamsError(err_msg)
 
     def extract_field(self, field):
         """ extract value from requests.Response.
         """
-        if not isinstance(field, basestring):
-            err_msg = u"Invalid extractor! => {}\n".format(field)
-            logger.log_error(err_msg)
+        if not isinstance(field, str):
+            err_msg = f"Invalid extractor! => {field}\n"
+            logger.error(err_msg)
             raise exceptions.ParamsError(err_msg)
 
-        msg = "extract: {}".format(field)
+        msg = f"extract: {field}"
 
         if field.startswith("$"):
             value = self._extract_field_with_jsonpath(field)
@@ -246,11 +268,8 @@ class ResponseObject(object):
         else:
             value = self._extract_field_with_delimiter(field)
 
-        if is_py2 and isinstance(value, unicode):
-            value = value.encode("utf-8")
-
-        msg += "\t=> {}".format(value)
-        logger.log_debug(msg)
+        msg += f"\t=> {value}"
+        logger.debug(msg)
 
         return value
 
@@ -274,7 +293,7 @@ class ResponseObject(object):
         if not extractors:
             return {}
 
-        logger.log_debug("start to extract from response object.")
+        logger.debug("start to extract from response object.")
         extracted_variables_mapping = OrderedDict()
         extract_binds_order_dict = utils.ensure_mapping_format(extractors)